requirements.txt
customtkinter
matplotlib
pillow
